+++ src/cli/cli-commands/run.ts
@@ -20,10 +20,9 @@ import { $registeredLlmToolsMessage } from '../../llm-providers/_common/register
 import type { PipelineJson } from '../../pipeline/PipelineJson/PipelineJson';
 import { $provideFilesystemForNode } from '../../scrapers/_common/register/$provideFilesystemForNode';
 import { $provideScrapersForNode } from '../../scrapers/_common/register/$provideScrapersForNode';
-import { scraperFetch } from '../../scrapers/_common/utils/scraperFetch';
+import { promptbookFetch } from '../../scrapers/_common/utils/promptbookFetch';
 import { JavascriptExecutionTools } from '../../scripting/javascript/JavascriptExecutionTools';
-import type { string_parameter_name } from '../../types/typeAliases';
-import type { string_parameter_value } from '../../types/typeAliases';
+import type { string_parameter_name, string_parameter_value } from '../../types/typeAliases';
 import { countLines } from '../../utils/expectation-counters/countLines';
 import { countWords } from '../../utils/expectation-counters/countWords';
 import { isFileExisting } from '../../utils/files/isFileExisting';
@@ -190,7 +189,7 @@ export function $initializeRunCommand(program: Program) {
             const tools = {
                 llm,
                 fs,
-                fetch: scraperFetch,
+                fetch: promptbookFetch,
                 scrapers: await $provideScrapersForNode({ fs, llm, executables }, prepareAndScrapeOptions),
                 script: [new JavascriptExecutionTools(cliOptions)],
             } satisfies ExecutionTools;