+++ src/cli/common/$provideLlmToolsForCli.ts
@@ -5,7 +5,7 @@ import { UnexpectedError } from '../../errors/UnexpectedError';
 import { $provideLlmToolsForWizzardOrCli } from '../../llm-providers/_common/register/$provideLlmToolsForWizzardOrCli';
 import type { CacheLlmToolsOptions } from '../../llm-providers/_common/utils/cache/CacheLlmToolsOptions';
 import type { LoginResponse } from '../../remote-server/types/RemoteServerOptions';
-import { scraperFetch } from '../../scrapers/_common/utils/scraperFetch';
+import { promptbookFetch } from '../../scrapers/_common/utils/promptbookFetch';
 import type { string_url } from '../../types/typeAliases';
 import type { really_unknown } from '../../utils/organization/really_unknown';
 import { TODO_USE } from '../../utils/organization/TODO_USE';
@@ -88,7 +88,7 @@ export function $provideLlmToolsForCli(options: ProvideLlmToolsForCliOptions) {
                 console.log('!!!', { loginUrl });
 
                 // TODO: [🧠] Should we use normal `fetch` or `scraperFetch`
-                const response = await scraperFetch(loginUrl, {
+                const response = await promptbookFetch(loginUrl, {
                     method: 'POST',
                     headers: {
                         'Content-Type': 'application/json',