+++ src/scrapers/_common/utils/makeKnowledgeSourceHandler.ts
@@ -4,10 +4,12 @@ import { dirname, join } from 'path';
 import spaceTrim from 'spacetrim';
 import type { SetOptional } from 'type-fest';
 import { knowledgeSourceContentToName } from '../../../commands/KNOWLEDGE/utils/knowledgeSourceContentToName';
-import { DEFAULT_DOWNLOAD_CACHE_DIRNAME } from '../../../config';
-import { DEFAULT_IS_VERBOSE } from '../../../config';
-import { DEFAULT_MAX_FILE_SIZE } from '../../../config';
-import { MAX_FILENAME_LENGTH } from '../../../config';
+import {
+    DEFAULT_DOWNLOAD_CACHE_DIRNAME,
+    DEFAULT_IS_VERBOSE,
+    DEFAULT_MAX_FILE_SIZE,
+    MAX_FILENAME_LENGTH,
+} from '../../../config';
 import { EnvironmentMismatchError } from '../../../errors/EnvironmentMismatchError';
 import { LimitReachedError } from '../../../errors/LimitReachedError';
 import { NotFoundError } from '../../../errors/NotFoundError';
@@ -25,7 +27,7 @@ import { TODO_USE } from '../../../utils/organization/TODO_USE';
 import { isValidFilePath } from '../../../utils/validators/filePath/isValidFilePath';
 import { isValidUrl } from '../../../utils/validators/url/isValidUrl';
 import type { ScraperSourceHandler } from '../Scraper';
-import { scraperFetch } from './scraperFetch';
+import { promptbookFetch } from './promptbookFetch';
 
 /**
  * @@@
@@ -39,7 +41,7 @@ export async function makeKnowledgeSourceHandler(
 ): Promise<ScraperSourceHandler> {
     // console.log('!! makeKnowledgeSourceHandler', knowledgeSource);
 
-    const { fetch = scraperFetch } = tools;
+    const { fetch = promptbookFetch } = tools;
     const { knowledgeSourceContent } = knowledgeSource;
     let { name } = knowledgeSource;
     const {